+++ cardano-node/src/Cardano/Node/Run.hs
@@ -391,7 +391,7 @@ handleSimpleNode runP p2pMode tracers nc onKernel = do
       EnabledP2PMode -> do
         traceWith (startupTracer tracers)
                   (StartupP2PInfo (ncDiffusionMode nc))
-        nt <- TopologyP2P.readTopologyFileOrError nc
+        nt <- TopologyP2P.readTopologyFileOrError (startupTracer tracers) nc
         let (localRoots, publicRoots) = producerAddresses nt
         traceWith (startupTracer tracers)
                 $ NetworkConfig localRoots
@@ -500,6 +500,7 @@ handleSimpleNode runP p2pMode tracers nc onKernel = do
                      developmentNtcVersions)
 
 #ifdef UNIX
+  -- only used when P2P is enabled
   updateTopologyConfiguration :: StrictTVar IO [(Int, Map RelayAccessPoint PeerAdvertise)]
                               -> StrictTVar IO [RelayAccessPoint]
                               -> StrictTVar IO UseLedgerAfter
@@ -507,7 +508,7 @@ handleSimpleNode runP p2pMode tracers nc onKernel = do
   updateTopologyConfiguration localRootsVar publicRootsVar useLedgerVar =
     Signals.Catch $ do
       traceWith (startupTracer tracers) NetworkConfigUpdate
-      result <- try $ readTopologyFileOrError nc
+      result <- try $ TopologyP2P.readTopologyFileOrError (startupTracer tracers) nc
       case result of
         Left (FatalError err) ->
           traceWith (startupTracer tracers)